[fix] Update post Oauth API
At first login, the refresh token was sent to the client as a cookie, and the refresh token and user information were stored in the database.
From the next login, the refresh token of the database will be delivered in cookies.
Due to the database schema change, the studylog and seed sql files were modified.